# Improve AvaloniaObject.GetValue performance