--- a/src/Avalonia.Base/DirectProperty.cs
+++ b/src/Avalonia.Base/DirectProperty.cs
@@ -97,7 +97,7 @@ namespace Avalonia
             metadata.Freeze();
 
             var result = new DirectProperty<TNewOwner, TValue>(
-                (DirectPropertyBase<TValue>)this,
+                this,
                 getter,
                 setter,
                 metadata);
@@ -144,9 +144,9 @@ namespace Avalonia
         }
 
         object? IDirectPropertyAccessor.GetUnsetValue(Type type)
-        {
-            var metadata = GetMetadata(type);
-            return metadata.UnsetValue;
-        }
+            => GetMetadata(type).UnsetValue;
+
+        object? IDirectPropertyAccessor.GetUnsetValue(AvaloniaObject owner)
+            => GetMetadata(owner).UnsetValue;
     }
 }