--- a/src/Avalonia.Base/PropertyStore/DirectBindingObserver.cs
+++ b/src/Avalonia.Base/PropertyStore/DirectBindingObserver.cs
@@ -15,7 +15,7 @@ namespace Avalonia.PropertyStore
         public DirectBindingObserver(ValueStore owner, DirectPropertyBase<T> property)
         {
             _owner = owner;
-            _hasDataValidation = property.GetMetadata(owner.Owner.GetType())?.EnableDataValidation ?? false;
+            _hasDataValidation = property.GetMetadata(owner.Owner).EnableDataValidation ?? false;
             Property = property;
         }