… src/Avalonia.Base/PropertyStore/EffectiveValue.cs
… src/Avalonia.Base/PropertyStore/EffectiveValue`1.cs
… src/Avalonia.Base/PropertyStore/LocalValueBindingObserverBase.cs
… src/Avalonia.Base/PropertyStore/SourceUntypedBindingEntry.cs
… src/Avalonia.Base/PropertyStore/TypedBindingEntry.cs
… src/Avalonia.Base/PropertyStore/ValueStore.cs
… src/Avalonia.Base/StyledProperty.cs
… src/Avalonia.Base/StyledPropertyMetadata`1.cs
… src/Avalonia.Base/Utilities/AvaloniaPropertyDictionary.cs
… src/Avalonia.Base/Utilities/TypeEqualityComparer.cs
… src/Markup/Avalonia.Markup.Xaml/MarkupExtensions/CompiledBindingExtension.cs
… src/Markup/Avalonia.Markup/Data/Binding.cs
… src/Markup/Avalonia.Markup/Data/BindingBase.cs
… src/Markup/Avalonia.Markup/Data/MultiBinding.cs
… src/Shared/StringCompatibilityExtensions.cs
… tests/Avalonia.Base.UnitTests/Utilities/AvaloniaPropertyDictionaryTests.cs
… tests/Avalonia.Benchmarks/ControlHierarchyCreator.cs