## What does the pull request do?
This PR improves the performance of `AvaloniaProperty.GetValue()`, which is used for every Avalonia property access.
This method is such a hot path that even tiny improvements can have a measurable effect.

## Numbers

Let's start with the `Remeasure` benchmark, which does a re-layout pass for ≈6500 nested `StackPanel`/`Buttons`:

### Before

|    Method |     Mean |     Error |    StdDev | Allocated |
|---------- |---------:|----------:|----------:|----------:|
| Remeasure | 4.638 ms | 0.0268 ms | 0.0250 ms |  76.86 KB |

### After

|    Method |     Mean |     Error |    StdDev |   Gen0 | Allocated |
|---------- |---------:|----------:|----------:|-------:|----------:|
| Remeasure | 2.843 ms | 0.0231 ms | 0.0216 ms | 3.9063 |  76.85 KB |

(Numbers are for Ryzen 9 5900X, Windows 11, x64)

Now for the micro-benchmarks:

### Before

|                             Method |        Mean |    Error |   StdDev |
|----------------------------------- |------------:|---------:|---------:|
|                   GetDefaultValues | 10,260.6 ns | 55.07 ns | 51.51 ns |
|                   Get_Local_Values |  4,219.3 ns | 33.76 ns | 31.58 ns |
| Get_Local_Values_With_Style_Values |  4,080.2 ns | 22.00 ns | 19.50 ns |

|           Method | PropertyCount |       Mean |     Error |    StdDev |
|----------------- |-------------- |-----------:|----------:|----------:|
| LookupProperties |             2 |   3.219 ns | 0.0590 ns | 0.0551 ns |
| LookupProperties |             6 |  14.549 ns | 0.1533 ns | 0.1359 ns |
| LookupProperties |            10 |  30.965 ns | 0.5022 ns | 0.4698 ns |
| LookupProperties |            20 |  77.665 ns | 0.3909 ns | 0.3656 ns |
| LookupProperties |            30 | 131.474 ns | 0.7703 ns | 0.7205 ns |

### After

|                             Method |       Mean |    Error |   StdDev |
|----------------------------------- |-----------:|---------:|---------:|
|                   GetDefaultValues | 1,178.9 ns |  4.35 ns |  4.07 ns |
|                   Get_Local_Values | 2,685.4 ns | 11.23 ns |  8.77 ns |
| Get_Local_Values_With_Style_Values | 2,710.9 ns | 19.54 ns | 18.27 ns |

|           Method | PropertyCount |      Mean |     Error |    StdDev |
|----------------- |-------------- |----------:|----------:|----------:|
| LookupProperties |             2 |  2.183 ns | 0.0150 ns | 0.0133 ns |
| LookupProperties |             6 | 10.633 ns | 0.0432 ns | 0.0383 ns |
| LookupProperties |            10 | 21.902 ns | 0.1005 ns | 0.0940 ns |
| LookupProperties |            20 | 56.363 ns | 0.3171 ns | 0.2967 ns |
| LookupProperties |            30 | 95.471 ns | 0.4494 ns | 0.4204 ns |

## How was the solution implemented (if it's not obvious)?

There are two main gains here:

### Default values

First, improving getting the default value was the main concern since that's the most common case. (Look at that almost 9x speedup!)

This is done by avoiding metadata lookup as much as possible. If we have only a single default value even for a property with multiple metadatas (`AddOwner()` calls), cache it in a field and return it asap.

If we still have several metadatas, lookup fast if possible. `Type.IsInstanceOfType(obj)` is about 25% faster for classes than `Type.IsAssignableFrom(obj.GetType())` (.NET 8). A dictionary is used only in last resort.

A `ReferenceEqualityComparer` is used for this dictionary, resulting in a ≈20% speedup per lookup.

`[MethodImpl(MethodImplOptions.AggressiveInlining)]` has been used where it made sense to do so (always measured first).

### Local values

Next was improving lookup for local values.

This is where micro-optimizations matter. `AvaloniaPropertyDictionary.TryGetValue()` is everywhere: every nanosecond counts here. Even if there's no local value for a property, we must always get through this path first. It has to be _fast_.

The binary search algorithm used here went through several iterations. I tried several versions of loop unrolling, unsafe code and SIMD, verifying performance numbers and checking the resulting x64 assembly code every time, ensuring there's no unneeded instruction. (The results may vary for other platforms, but I still expect a gain.)

It turns out that a very simple loop is still the fastest (see the `LookupProperties` benchmark above), with some extras such as a forced bounds check removal (that couldn't be elided naturally by the JIT). It's fast enough that the linear search part has been removed.

The binary search in `FindEntry` has been copied manually in `TryGetValue`, since even aggressive inlining wasn't enough to squeeze that last drop of performance out of the method.

The property's `Id` is stored directly inside the `Entry`, avoiding an indirection and probably allowing everything to fit inside a CPU cache line (the change was measurable in the `LookupProperties` benchmark).

### Misc

Used `ArgumentNullException.ThrowIfNull` where possible, ensuring the `throw` part is never inlined, allowing the JIT to inline the caller if it chooses to do so.

Made `Optional<T>.GetValueOrDefault()` an unconditonal single field access, since `_value` will be `default` if there's no value (`Nullable<T>` has a similar implementation).